oneview-monasca/requirements.txt
configparser>=3.5.0
future>=0.15.2
ujson
hpOneView
kazoo
keystoneauth1>=2.8.0
pbr>=1.10.0
pika>=0.10.0
python-dateutil>=2.5.3
python-ironicclient>=1.3.1
python-keystoneclient>=3.1.0
python-monascaclient>=1.0.30
pytz
six>=1.10.0
stevedore>=1.14.0
tooz
tzlocal
retrying
python-oneviewclient
